User verification, Admin

+++ routers/todos.py
@@ -7,13 +7,11 @@ from Model.ToDoRequest import *
 from database import engine, SessionLocal
 from starlette import status
 from pydantic import BaseModel, Field
-
-
+from .auth import get_current_user
 
 router = APIRouter(
     tags=['ToDo']
 )
-
 
 def get_db():
     db = SessionLocal()
@@ -22,18 +20,23 @@ def get_db():
     finally:
         db.close()
 
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
-#get all Todos
+# get all Todos
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: Session = Depends(get_db)):
-    return db.query(ToDo).all()
+async def read_all(user: user_dependency, db: Session = Depends(get_db)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    return db.query(ToDo).filter(ToDo.owner_id == user.get('id')).all()
 
 
 
 #get single Todo
 @router.get("/todo/{id}", status_code=status.HTTP_200_OK)
-async def get_todo(id: int = Path(gt=0), db: Session = Depends(get_db)):
-    todo = db.query(ToDo).filter(ToDo.id == id).first()
+async def get_todo(user: user_dependency, id: int = Path(gt=0), db: Session = Depends(get_db)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    todo = db.query(ToDo).filter(ToDo.id == id).filter(ToDo.owner_id == user.get('id')).first()
     if todo is not None:
         return todo
     raise HTTPException(status_code=404, detail='Not found')
@@ -41,8 +44,10 @@ async def get_todo(id: int = Path(gt=0), db: Session = Depends(get_db)):
 
 #create new Todo
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_new(request: ToDoRequest, db: Session = Depends(get_db)):
-    todo = ToDo(**request.dict())
+async def create_new(request: ToDoRequest, user: user_dependency, db: Session = Depends(get_db)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    todo = ToDo(**request.dict(), owner_id = user.get('id'))
     db.add(todo)
     db.commit()
     return HTTPException(status_code=200, detail='Created', headers='')
@@ -50,8 +55,12 @@ async def create_new(request: ToDoRequest, db: Session = Depends(get_db)):
 
 #update Todo
 @router.put("/todo/{id}", status_code=status.HTTP_200_OK)
-async def update_todo(id: int, request: ToDoRequest, db: Session = Depends(get_db)):
-    todo = db.query(ToDo).filter(ToDo.id == id).first()
+async def update_todo(user: user_dependency, request: ToDoRequest, id: int = Path(gt=0), db: Session = Depends(get_db)):
+
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+
+    todo = db.query(ToDo).filter(ToDo.id == id).filter(ToDo.owner_id == user.get('id')).first()
 
     if todo is None:
         raise HTTPException(status_code=404, detail='Not found')
@@ -59,22 +68,25 @@ async def update_todo(id: int, request: ToDoRequest, db: Session = Depends(get_d
     todo.title = request.title
     todo.description = request.description
     todo.priority = request.priority
-    todo.complite = request.complite
+    todo.complete = request.complete
 
     db.add(todo)
     db.commit()
     return HTTPException(status_code=200, detail='Updated', headers='')
 
 
-#delete todo
+#delete Todo
 @router.delete("/delete/{id}", status_code=status.HTTP_200_OK)
-async def delete_todo(id:int = Path(gt=0), db: Session = Depends(get_db)):
-    todo = db.query(ToDo).filter(ToDo.id == id).first()
+async def delete_todo(user: user_dependency, id:int = Path(gt=0), db: Session = Depends(get_db)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+
+    todo = db.query(ToDo).filter(ToDo.id == id).filter(ToDo.owner_id == user.get('id')).first()
 
     if todo is None:
         raise HTTPException(status_code=404, detail='Not found')
 
-    db.query(ToDo).filter(ToDo.id == id).delete()
+    db.query(ToDo).filter(ToDo.id == id).filter(ToDo.owner_id == user.get('id')).delete()
     db.commit()
     return HTTPException(status_code=200, detail='Deleted', headers='')
